requirements.txt
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.17.0
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-socketio==5.3.6
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
mistune==3.1.2
packaging==24.2
platformdirs==4.3.6
PyMySQL==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
virtualenv==20.29.2
Werkzeug==3.1.3
python-socketio==5.10.0
python-engineio==4.8.0

PyJWT==2.8.0       
python-dotenv==1.0.1
gunicorn==21.2.0
flask-migrate==4.0.7
gunicorn==21.2.0
